requirements.txt
Flask==2.0.1
boto3==1.17.86
requests==2.25.0
pycognito==2021.3.1
pymongo==3.11.4
python-dotenv==0.17.1